Moved logic for benchmarking to examples/tti_benchmark.py
